+++ backend/engines/api/app/serializers/spree/api/v3/order_serializer.rb
@@ -15,6 +15,7 @@ module Spree
                  included_tax_total: :string, display_included_tax_total: :string,
                  additional_tax_total: :string, display_additional_tax_total: :string,
                  total: :string, display_total: :string, completed_at: 'string | null',
+                 bill_address: 'StoreAddress | null', ship_address: 'StoreAddress | null',
                  payment_methods: 'StorePaymentMethod[]'
 
         attributes :number, :state, :token, :email, :special_instructions,
@@ -25,29 +26,12 @@ module Spree
                    :additional_tax_total, :display_additional_tax_total, :total, :display_total,
                    completed_at: :iso8601, created_at: :iso8601, updated_at: :iso8601
 
-        many :order_promotions,
-             resource: Spree.api.order_promotion_serializer,
-             if: proc { params[:includes]&.include?('order_promotions') }
-
-        many :line_items,
-             resource: Spree.api.line_item_serializer,
-             if: proc { params[:includes]&.include?('line_items') }
-
-        many :shipments,
-             resource: Spree.api.shipment_serializer,
-             if: proc { params[:includes]&.include?('shipments') }
-
-        many :payments,
-             resource: Spree.api.payment_serializer,
-             if: proc { params[:includes]&.include?('payments') }
-
-        one :bill_address,
-            resource: Spree.api.address_serializer,
-            if: proc { params[:includes]&.include?('bill_address') }
-
-        one :ship_address,
-            resource: Spree.api.address_serializer,
-            if: proc { params[:includes]&.include?('ship_address') }
+        many :order_promotions, resource: Spree.api.order_promotion_serializer
+        many :line_items, resource: Spree.api.line_item_serializer
+        many :shipments, resource: Spree.api.shipment_serializer
+        many :payments, resource: Spree.api.payment_serializer
+        one :bill_address, resource: Spree.api.address_serializer
+        one :ship_address, resource: Spree.api.address_serializer
 
         many :payment_methods, resource: Spree.api.payment_method_serializer
       end